refactor: move tweets snapshot logic to the useTweetsSnapshot hook
