+++ src/components/Feed/Feed.tsx
@@ -1,44 +1,24 @@
-import React, { useEffect, useState } from 'react';
-import { collection, onSnapshot, orderBy, query } from 'firebase/firestore';
+import React, { useMemo } from 'react';
+import { collection, orderBy, query } from 'firebase/firestore';
 import { Head } from '@/pages/HomePage/HomePage.styled';
 import { Switch } from '@/components/ui/Switch/Switch';
 import { NewTweet } from '@/components/ui/NewTweet/NewTweet';
 import { Tweets } from '@/components/ui/Tweets/Tweets';
-import { Tweet, Tweet as TweetType } from '@/types/tweet';
 import { User } from '@/types/user';
 import { db } from '@/db/firesbase';
 import { Collections } from '@/types/collections';
+import { useTweetsSnapshot } from '@/hooks/useTweetsSnapshot';
 
 interface FeedProps {
   user: User;
 }
 
 const Feed = ({ user }: FeedProps) => {
-  const [tweets, setTweets] = useState<Tweet[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    const tweetsQuery = query(
-      collection(db, Collections.Tweets),
-      orderBy('date', 'desc'),
-    );
-
-    setIsLoading(true);
-    const unsubscribe = onSnapshot(tweetsQuery, (snapshot) => {
-      const updatedTweets = snapshot.docs.map((tweetDoc) => {
-        const tweetData = tweetDoc.data();
-        return {
-          id: tweetDoc.id,
-          ...tweetData,
-        };
-      }) as TweetType[];
-
-      setIsLoading(false);
-      setTweets(updatedTweets);
-    });
-
-    return unsubscribe;
-  }, [user.id]);
+  const tweetsQuery = useMemo(
+    () => query(collection(db, Collections.Tweets), orderBy('date', 'desc')),
+    [],
+  );
+  const { tweets, isTweetsLoading } = useTweetsSnapshot(tweetsQuery);
 
   return (
     <>
@@ -47,7 +27,7 @@ const Feed = ({ user }: FeedProps) => {
         <Switch onChange={() => {}} />
       </Head>
       <NewTweet user={user} />
-      <Tweets tweets={tweets} isLoading={isLoading} user={user} />
+      <Tweets tweets={tweets} isLoading={isTweetsLoading} user={user} />
     </>
   );
 };
